Updated stricter uint64 and uint32 types

+++ timeline.go
@@ -1,11 +1,11 @@
 package mpd
 
 type TimeLine struct {
-	Start int
-	End   int
+	Start uint64
+	End   uint64
 }
 
-func NewTimeLine(start, end int) TimeLine {
+func NewTimeLine(start, end uint64) TimeLine {
 	return TimeLine{
 		Start: start,
 		End:   end,
